Fix compilation on clang 8.0 and modern ubuntu disco Most issues were around stale ffmpeg. Build originally worked well on GCC but at some point it also discovered the same issues.

+++ src/audio/renderers/OpenAL/OpenALRenderer.cpp
@@ -171,7 +171,7 @@ namespace Audio {
                     clearAlError();
                     
                     ALCint params[] = {
-                        ALC_FREQUENCY, format.sampleFrequency,
+                        ALC_FREQUENCY, static_cast<int>(format.sampleFrequency),
                         0
                     };
