+++ src/gfx/nav/navitemstodraw.h
@@ -28,11 +28,11 @@ void NavigationSystem::DrawCursor( float x, float y, float wid, float hei, const
             x,          y,          0,
             x,          y-hei,      0,
             x,          y,          0,
-            x+wid,      y-0.75*hei, 0,
+            x+wid,      static_cast<float>(y-0.75*hei), 0,
             x,          y-hei,      0,
-            x+0.35*wid, y-0.6*hei,  0,
-            x+0.35*wid, y-0.6*hei,  0,
-            x+wid,      y-0.75*hei, 0,
+            static_cast<float>(x+0.35*wid), static_cast<float>(y-0.6*hei),  0,
+            static_cast<float>(x+0.35*wid), static_cast<float>(y-0.6*hei),  0,
+            x+wid,      static_cast<float>(y-0.75*hei), 0,
         };
         GFXDraw( GFXLINE, verts, 8 );
